pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-forge"
version = "0.1.0"
description = "Forge router — directs agents to the right Eidos forge"
requires-python = ">=3.11"
license = "MIT"
authors = [{name = "Daniel Shanklin", email = "daniel@eidosagi.com"}]
dependencies = [
    "mcp>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
forge-forge-mcp = "mcp_server:main"

[tool.setuptools]
py-modules = ["mcp_server"]
